feat(packages): add new features to packages entry (#10)
* feat(packages): allow pkgs from flake.nix file

There are situations where authors want to create packages on the fly
without having to create a new file in the packages directory. nixDir
now provides an easy way to pass packages in the flake.nix file.

* feat(packages): generate all package

When enabling the `generateAllPackage` option, nixDir will create a new
package that depends on all packages from the flake. This is useful if
we want to build all the packages to later share them with a remote
nix-store.

* docs: include new packages features on README

* ci: add nixt and bats tests to pre-commit checks