fix: packages entry creation check
Previously, we would only create the flake's packages entry only when
there was a `nix/packages` directory.

This approach was inaccurate, as we have many other ways to generate
packages (when using either the `packages` or `generatelAllPackage`
arguments in the `buildFlake` call).

This change accurately creates the flake's `packages` entry when any of
the packages specifications are used.